api/requirements.txt
flask
flask-cors
tensorflow
pillow
numpy
requests